requirements.txt
openai>=1.10.0
pymongo>=4.6.1
python-dotenv>=1.0.1
tqdm>=4.66.2  # optional, for loading indicators
